Move files to blog folder
